refactor importing of product components to use index.js
