+++ client/components/products/index.js
@@ -1,0 +1,2 @@
+export {default as AllProducts} from './AllProducts'
+export {default as ProductDetail} from './ProductDetail'
